+++ parlai/agents/transformer/transformer.py
@@ -10,6 +10,8 @@ from parlai.utils.torch import padded_3d
 from parlai.core.torch_classifier_agent import TorchClassifierAgent
 from parlai.core.torch_ranker_agent import TorchRankerAgent
 from parlai.core.torch_generator_agent import TorchGeneratorAgent
+from parlai.utils.misc import recursive_getattr
+from parlai.utils.logging import logging
 
 from .modules import (
     TransformerMemNetModel,
@@ -325,6 +327,34 @@ class TransformerGeneratorAgent(TorchGeneratorAgent):
                 model.encoder.embeddings.weight, self.opt['embedding_type']
             )
         return model
+
+    def _resize_token_embeddings(self, state_dict, msg=None):
+        """
+        Resize the token embeddings when are adding extra special tokens.
+        """
+        # map extra special tokens carefully
+        new_size = self.model.embeddings.weight.size()[0]
+        orig_size = state_dict['embeddings.weight'].size()[0]
+        logging.info(f'Resizing token embeddings from {orig_size} to {new_size}')
+        if new_size <= orig_size:
+            # new size should be greater than original size,
+            # as we are adding special tokens
+            raise RuntimeError(msg)
+
+        for emb_weights in [
+            'embeddings.weight',
+            'encoder.embeddings.weight',
+            'decoder.embeddings.weight',
+        ]:
+            # get new_embs
+            old_embs = state_dict[emb_weights]
+            new_embs = recursive_getattr(self.model, emb_weights).to(old_embs.device)
+            # copy over old weights
+            new_embs.data[:orig_size, :] = old_embs.data[:orig_size, :]
+            # reset in state dict
+            state_dict[emb_weights] = new_embs
+
+        return state_dict
 
 
 class TransformerClassifierAgent(TorchClassifierAgent):
