+++ parlai/core/torch_agent.py
@@ -653,6 +653,12 @@ class TorchAgent(ABC, Agent):
             choices=[None, 'end'],
             help='Add special token to the end of history encoding.',
         )
+        agent.add_argument(
+            '--special-tok-lst',
+            type=str,
+            default=None,
+            help='Comma separated list of special tokens',
+        )
         # GPU arguments
         # these gpu options are all mutually exclusive, and should error if the
         # user tries to present multiple of them
@@ -801,10 +807,27 @@ class TorchAgent(ABC, Agent):
         place to do it.
         """
         d = self.dictionary_class()(self.opt)
+        self.special_toks = self._get_special_tokens()
+        if self.special_toks:
+            d.add_extra_special_tokens(self.special_toks)
+
         if self.opt.get('person_tokens'):
             d[self.P1_TOKEN] = 999_999_999
             d[self.P2_TOKEN] = 999_999_998
         return d
+
+    def _resize_token_embeddings(self, state_dict, msg=None):
+        """
+        Must define this for your agent if you wish to add additional special tokens.
+
+        Must make a call to resize the token embeddings and load the model state dict
+        with the resized token embeddings.
+        """
+        raise RuntimeError(
+            'If you are intending to add special tokens to an already pretrained model, '
+            'you must write the function `_resize_token_embeddings` for your specific '
+            'agent.'
+        )
 
     def _get_init_model(self, opt: Opt, shared):
         """
@@ -845,6 +868,16 @@ class TorchAgent(ABC, Agent):
 
         return init_model, is_finetune
 
+    def _get_special_tokens(self) -> List[str]:
+        """
+        Return list of special tokens.
+
+        Made easily overridable for special cases.
+        """
+        if self.opt.get('special_tok_lst') is not None:
+            return self.opt['special_tok_lst'].split(',')
+        return []
+
     @abstractmethod
     def build_model(self):
         """
@@ -878,6 +911,10 @@ class TorchAgent(ABC, Agent):
             type of optimizer being loaded, if changed will skip loading
             optimizer states
         """
+        if hasattr(self, 'resized') and self.resized:
+            optim_states = None
+            logging.warn('Not loading optimizer due to resize in token embeddings')
+
         opt = self.opt
 
         # set up optimizer args
@@ -1810,14 +1847,19 @@ class TorchAgent(ABC, Agent):
         except RuntimeError as msg:
             msg_ = str(msg)
             if 'size mismatch' in msg_ and 'embedding' in msg_:
-                raise RuntimeError(
-                    f'{msg_}\n'
-                    '-----------------\n'
-                    'Could not load the model due to a size mismatch in the '
-                    'embeddings. A common reason for this is trying to load '
-                    'a model trained with fp16 but loaded without fp16. Try '
-                    'adding --fp16 true or --force-fp16-tokens true.'
-                )
+                if hasattr(self, 'special_toks') and len(self.special_toks) > 0:
+                    state_dict = self._resize_token_embeddings(state_dict, msg_)
+                    self.model.load_state_dict(state_dict)
+                    self.resized = True  # make note that we resized here
+                else:
+                    raise RuntimeError(
+                        f'{msg_}\n'
+                        '-----------------\n'
+                        'Could not load the model due to a size mismatch in the '
+                        'embeddings. A common reason for this is trying to load '
+                        'a model trained with fp16 but loaded without fp16. Try '
+                        'adding --fp16 true or --force-fp16-tokens true.'
+                    )
             else:
                 raise
